// app/models/Account.ts
export class Account {
    mobileNumber: string;
    email: string;
    firstName: string;
    lastName: string; 
    customerIdentifier: number; 
    accountNumber: number;
    password: string;
    creationDate: Date;
    fee: number;
    managerUpdatedFee:number;
    isManagerAccount: boolean; 
}